+++ gradle-plugin/src/main/kotlin/dev/zacsweers/metro/gradle/MetroGradleSubplugin.kt
@@ -161,6 +161,7 @@ public class MetroGradleSubplugin : KotlinCompilerPluginSupportPlugin {
         add(lazyOption("transform-providers-to-private", extension.transformProvidersToPrivate))
         add(lazyOption("shrink-unused-bindings", extension.shrinkUnusedBindings))
         add(lazyOption("chunk-field-inits", extension.chunkFieldInits))
+        add(lazyOption("statements-per-init-fun", extension.statementsPerInitFun))
         add(lazyOption("public-provider-severity", extension.publicProviderSeverity))
         add(
           lazyOption(